MDEV-14241: Server crash in key_copy / get_matching_chain_by_join_key or valgrind warnings
In this case we were using the optimization derived_with_keys but we could not create a key because
the length of the key was greater than the max allowed(MI_MAX_KEY_LENGTH). To do the join we needed
to create a hash join key instead, but in the explain output it showed that we were still referring
to derived keys which were created but not used.