MDEV-15016: multiple page cleaner threads use a lot of CPU on idle server
Problem was that when number of page cleaner threads was increased
we sent a event is_requested indicating there is "work" to be done
when there was no work to be done and workers were left on busy loop.
When number of page cleaner threads is decreased we need to sent
is_requested to indicate that there is "work" to be done, but we
must also reset it to avoid remaining threads think there is real
work for them also.

buf_flush_set_page_cleaner_thread_cnt
	When number of page cleaner threads is increasing we do
	not need to wake existing workers. If number of threads is
	decreasing, we need to wake them to  let them shutdown.

buf_flush_page_cleaner_worker
	When page cleaner worker thread is shutting down because
	number of page cleaner threads is decreased we need to
	reset is_requested event as it is used to indicate
	there is "work" to do.

Manually tested on both (1) idle system by increasing and decreasing
number of page cleaner threads and (2) on system with sysbench
workload by increasing and decreasing number of page cleaner threads
and monitoring mysqld CPU usage